# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: AutiGuard API Spec
  description: Production-grade, AWS-native API for AI-powered safety and emotional monitoring of individuals with Autism Spectrum Disorder (ASD).
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Dev Server
paths:
  /auth/login:
    post:
      summary: User Login
      description: Authenticates with Cognito User Pool and returns JWT tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successful Authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  id_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          description: Invalid Credentials

  /wearers:
    get:
      summary: List Wearers
      description: Retrieve wearer profiles associated with the caregiver's organization.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of wearers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WearerProfile'
    post:
      summary: Create Wearer Profile
      description: Adds a new individual with ASD to the platform. Enforces KMS encryption on PII fields.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WearerCreate'
      responses:
        '201':
          description: Wearer profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WearerProfile'

  /wearers/{id}:
    get:
      summary: Get Wearer Profile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wearer Profile detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WearerProfile'
    put:
      summary: Update Wearer Profile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WearerCreate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WearerProfile'

  /telemetry:
    post:
      summary: Ingest Telemetry Batch
      description: Ingests wearable metrics (Heart Rate, GPS, Accelerometer, battery). Can be invoked via IoT Core Rule.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryBatch'
      responses:
        '202':
          description: Telemetry data accepted for processing

  /wearers/{id}/telemetry:
    get:
      summary: Query Wearer Telemetry
      description: Returns timeseries data for visual charts.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: range
          in: query
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d]
            default: 24h
      responses:
        '200':
          description: Historical telemetry array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryItem'

  /alerts:
    get:
      summary: List Org Alerts
      description: Get active and past alert feeds across all wearers in the organization.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Alert feeds list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertItem'

  /alerts/{id}/acknowledge:
    post:
      summary: Acknowledge Alert
      description: Changes alert status to acknowledged, recording user details to audit trial.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert acknowledged successfully

  /geofence:
    post:
      summary: Create Geofence safe zone
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceCreate'
      responses:
        '201':
          description: Geofence created

  /qr/{wearerId}/rotate:
    post:
      summary: Rotate QR Token
      description: Explicitly rotates the signed KMS ephemeral QR access token.
      security:
        - BearerAuth: []
      parameters:
        - name: wearerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New signed QR payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr_payload:
                    type: string
                  expires_at:
                    type: string
                    format: date-time

  /qr/resolve/{token}:
    get:
      summary: Resolve QR Code Token
      description: Public rate-limited resolution. Tiered data access depending on JWT role or scanner relationship.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tiered Wearer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRTieredResponse'

  /ml/stress-score:
    post:
      summary: SageMaker Stress Inference
      description: Evaluates features vector to calculate stress index (0-100).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hr_variability
                - motion_intensity
                - time_of_day
              properties:
                hr_variability:
                  type: number
                motion_intensity:
                  type: number
                time_of_day:
                  type: string
      responses:
        '200':
          description: Predicted stress index
          content:
            application/json:
              schema:
                type: object
                properties:
                  stress_index:
                    type: integer
                  risk_level:
                    type: string
                    enum: [NORMAL, ELEVATED, SEVERE]

  /comms:
    post:
      summary: Ingest Non-Verbal Communication Log
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wearer_id
                - category_code
              properties:
                wearer_id:
                  type: string
                category_code:
                  type: string
                  enum: [HUNGER, RESTROOM, DISCOMFORT, FEAR, ANXIETY]
      responses:
        '201':
          description: Event stored in logs
    get:
      summary: Retrieve Non-verbal Logs
      security:
        - BearerAuth: []
      parameters:
        - name: wearer_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of communication items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommLogItem'

  /reports/{wearerId}:
    get:
      summary: Generate Reports
      security:
        - BearerAuth: []
      parameters:
        - name: wearerId
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, monthly]
            default: weekly
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, csv]
            default: pdf
      responses:
        '200':
          description: Exported Report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    WearerCreate:
      type: object
      required:
        - first_name
        - last_name
        - dob
      properties:
        first_name:
          type: string
        last_name:
          type: string
        dob:
          type: string
          format: date
        medical_notes:
          type: string
        allergies:
          type: string
        medications:
          type: string
        qr_tiering_rules:
          type: object

    WearerProfile:
      type: object
      properties:
        wearer_id:
          type: string
        org_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        dob:
          type: string
        medical_notes:
          type: string
        allergies:
          type: string
        medications:
          type: string
        qr_payload:
          type: string
        created_at:
          type: string
          format: date-time

    TelemetryBatch:
      type: object
      required:
        - wearer_id
        - readings
      properties:
        wearer_id:
          type: string
        readings:
          type: array
          items:
            $ref: '#/components/schemas/TelemetryItem'

    TelemetryItem:
      type: object
      required:
        - timestamp
      properties:
        timestamp:
          type: string
          format: date-time
        heart_rate:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        accel_x:
          type: number
        accel_y:
          type: number
        accel_z:
          type: number
        accel_magnitude:
          type: number
        stress_index:
          type: number
        risk_level:
          type: string
        battery_level:
          type: number
        connectivity_status:
          type: string

    AlertItem:
      type: object
      properties:
        alert_id:
          type: string
        wearer_id:
          type: string
        wearer_name:
          type: string
        type:
          type: string
        severity:
          type: string
        details:
          type: object
        ack_status:
          type: string
        timestamp:
          type: string
          format: date-time

    GeofenceCreate:
      type: object
      required:
        - wearer_id
        - name
        - type
        - coordinates
      properties:
        wearer_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [radius, polygon]
        coordinates:
          type: array
          items:
            type: object
            properties:
              lat:
                type: number
              lng:
                type: number
        radius_meters:
          type: number

    CommLogItem:
      type: object
      properties:
        event_id:
          type: string
        category_code:
          type: string
        timestamp:
          type: string
          format: date-time

    QRTieredResponse:
      type: object
      properties:
        tier:
          type: integer
          enum: [1, 2, 3]
        first_name:
          type: string
        public_message:
          type: string
        emergency_contact:
          type: string
        medical_notes:
          type: string
        allergies:
          type: string
        medications:
          type: string